requirements.txt
APScheduler==3.6.3
astroid==2.5.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.7.4
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.3
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
